face-cutter/requirements.txt
Pillow
boto3
configparser
requests